p4i integration Integrate 104405 from 3 to main: Integrate 104404 from 3.2 to 3: Integrate 104403 from 3.1 to 3.2: warnings are there for a reason
[git-p4: depot-paths = "//depot/qt/main/": change = 104406]

+++ src/table/qtable.cpp
@@ -5303,9 +5303,9 @@ void QTable::activateNextCell()
 	 ( currentSel->leftCol() == currentSel->rightCol() &&
 	   currentSel->topRow() == currentSel->bottomRow() ) ) {
 	clearSelection();
-	if ( curRow < numRows() - d->hiddenRows.count() - 1 )
+	if ( (uint)curRow < numRows() - d->hiddenRows.count() - 1 )
 	    setCurrentCell( curRow + 1, curCol );
-	else if ( curCol < numCols() - d->hiddenCols.count() - 1 )
+	else if ( (uint)curCol < numCols() - d->hiddenCols.count() - 1 )
 	    setCurrentCell( 0, curCol + 1 );
 	else
 	    setCurrentCell( 0, 0 );
